tests/wpt/web-platform-tests/web-nfc/NDEFReadingEvent_constructor.https.html

<!DOCTYPE html>
<title>NDEFReadingEvent constructor</title>
<link rel="help" href="https://w3c.github.io/web-nfc/#dom-ndefreadingevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/nfc-helpers.js"></script>
<script>

  test(() => {
    assert_equals(NDEFReadingEvent.length, 2);
    assert_throws(new TypeError, () => new NDEFReadingEvent('message'));
  }, 'NDEFReadingEvent constructor without init dict');

  test(() => {
    const message = createMessage([createJsonRecord(test_json_data)]);
    const event = new NDEFReadingEvent('type', {serialNumber: null, message: message});
    assert_equals(event.serialNumber, '', 'serialNumber');
  }, 'NDEFReadingEvent constructor with null serialNumber');

  test(() => {
    const message = createMessage([createJsonRecord(test_json_data)]);
    const event = new NDEFReadingEvent('type', {message: message});
    assert_equals(event.serialNumber, '', 'serialNumber');
  }, 'NDEFReadingEvent constructor with serialNumber not present');

  test(() => {
    const event = new NDEFReadingEvent('type', {serialNumber: '', message: null});
    assert_equals(0, event.message.records.length, 'no records');
    assert_equals(0, event.message.url.length, 'empty url');
  }, 'NDEFReadingEvent constructor with null message');

  test(() => {
    const message = createMessage([createJsonRecord(test_json_data)]);
    const event = new NDEFReadingEvent('type', {serialNumber: '', message: message});
    assert_equals(event.type, 'type', 'type');
    assert_equals(event.serialNumber, '', 'serialNumber');
    assertWebNDEFMessagesEqual(event.message, new NDEFMessage(message));
  }, 'NDEFReadingEvent constructor with valid parameters');

</script>
